refactor: separate modules for uninstall and custom field utils
(cherry picked from commit 089204ab08fc37aa870a0bd59a900d3ad2d093c9)
