+++ india_compliance/gst_india/doctype/gst_settings/gst_settings.py
@@ -16,7 +16,8 @@ from india_compliance.gst_india.page.india_compliance_account import (
     _disable_api_promo,
     post_login,
 )
-from india_compliance.gst_india.utils import can_enable_api, toggle_custom_fields
+from india_compliance.gst_india.utils import can_enable_api
+from india_compliance.gst_india.utils.custom_fields import toggle_custom_fields
 
 
 class GSTSettings(Document):
@@ -56,7 +57,6 @@ class GSTSettings(Document):
         company_wise_account_types = {}
 
         for row in self.gst_accounts:
-
             # Validate Duplicate Accounts
             for fieldname in GST_ACCOUNT_FIELDS:
                 account = row.get(fieldname)
