+++ india_compliance/gst_india/setup/__init__.py
@@ -13,10 +13,8 @@ from india_compliance.gst_india.constants.custom_fields import (
     SALES_REVERSE_CHARGE_FIELDS,
 )
 from india_compliance.gst_india.setup.property_setters import get_property_setters
-from india_compliance.gst_india.utils import get_data_file_path, toggle_custom_fields
-from india_compliance.patches.post_install.update_e_invoice_fields_and_logs import (
-    delete_custom_fields as _delete_custom_fields,
-)
+from india_compliance.gst_india.utils import get_data_file_path
+from india_compliance.gst_india.utils.custom_fields import toggle_custom_fields
 
 ITEM_VARIANT_FIELDNAMES = frozenset(("gst_hsn_code", "is_nil_exempt", "is_non_gst"))
 
@@ -31,49 +29,16 @@ def after_install():
     add_fields_to_item_variant_settings()
 
 
-def before_uninstall():
-    delete_custom_fields()
-    delete_property_setters()
-    remove_fields_from_item_variant_settings()
-
-
 def create_custom_fields():
     # Validation ignored for faster creation
     # Will not fail if a core field with same name already exists (!)
     # Will update a custom field if it already exists
-    _create_custom_fields(
-        _get_custom_fields_map(
-            CUSTOM_FIELDS,
-            SALES_REVERSE_CHARGE_FIELDS,
-            E_INVOICE_FIELDS,
-            E_WAYBILL_FIELDS,
-        ),
-        ignore_validate=True,
-    )
+    _create_custom_fields(get_all_custom_fields(), ignore_validate=True)
 
 
 def create_property_setters():
     for property_setter in get_property_setters():
         frappe.make_property_setter(property_setter)
-
-
-def delete_custom_fields():
-    _delete_custom_fields(
-        _get_custom_fields_map(
-            CUSTOM_FIELDS,
-            SALES_REVERSE_CHARGE_FIELDS,
-            E_INVOICE_FIELDS,
-            E_WAYBILL_FIELDS,
-        )
-    )
-
-
-def delete_property_setters():
-    for property_setter in get_property_setters():
-        keys_to_update = ["doc_type", "field_name", "property", "value"]
-        # Update the keys to match with the property setter fields
-        filters = dict(zip(keys_to_update, list(property_setter.values())))
-        frappe.db.delete("Property Setter", filters)
 
 
 def create_address_template():
@@ -224,16 +189,23 @@ def show_accounts_settings_override_warning():
     )
 
     click.secho(
-        "This is being set as Billing Address, since that's the correct "
-        "address for determining GST applicablility.",
+        (
+            "This is being set as Billing Address, since that's the correct "
+            "address for determining GST applicablility."
+        ),
         fg="yellow",
     )
 
 
-def _get_custom_fields_map(*custom_fields_list):
+def get_all_custom_fields():
     result = {}
 
-    for custom_fields in custom_fields_list:
+    for custom_fields in (
+        CUSTOM_FIELDS,
+        SALES_REVERSE_CHARGE_FIELDS,
+        E_INVOICE_FIELDS,
+        E_WAYBILL_FIELDS,
+    ):
         for doctypes, fields in custom_fields.items():
             if isinstance(fields, dict):
                 fields = [fields]
@@ -241,11 +213,3 @@ def _get_custom_fields_map(*custom_fields_list):
             result.setdefault(doctypes, []).extend(fields)
 
     return result
-
-
-def remove_fields_from_item_variant_settings():
-    settings = frappe.get_doc("Item Variant Settings")
-    settings.fields = [
-        row for row in settings.fields if row.field_name not in ITEM_VARIANT_FIELDNAMES
-    ]
-    settings.save()
