+++ india_compliance/gst_india/utils/__init__.py
@@ -318,37 +318,6 @@ def get_all_gst_accounts(company):
     return accounts_list
 
 
-def toggle_custom_fields(custom_fields, show):
-    """
-    Show / hide custom fields
-
-    :param custom_fields: a dict like `{'Sales Invoice': [{fieldname: 'test', ...}]}`
-    :param show: True to show fields, False to hide
-    """
-
-    for doctypes, fields in custom_fields.items():
-        if isinstance(fields, dict):
-            # only one field
-            fields = [fields]
-
-        if isinstance(doctypes, str):
-            # only one doctype
-            doctypes = (doctypes,)
-
-        for doctype in doctypes:
-            frappe.db.set_value(
-                "Custom Field",
-                {
-                    "dt": doctype,
-                    "fieldname": ["in", [field["fieldname"] for field in fields]],
-                },
-                "hidden",
-                int(not show),
-            )
-
-            frappe.clear_cache(doctype=doctype)
-
-
 def parse_datetime(value, day_first=False):
     """Convert IST string to offset-naive system time"""
 
@@ -405,22 +374,6 @@ def get_titlecase_version(word, all_caps=False, **kwargs):
 
     if word in ABBREVIATIONS:
         return word
-
-
-def delete_old_fields(fields, doctypes):
-    if isinstance(fields, str):
-        fields = (fields,)
-
-    if isinstance(doctypes, str):
-        doctypes = (doctypes,)
-
-    frappe.db.delete(
-        "Custom Field",
-        {
-            "fieldname": ("in", fields),
-            "dt": ("in", doctypes),
-        },
-    )
 
 
 def is_api_enabled(settings=None):
