+++ india_compliance/patches/post_install/set_default_gst_settings.py
@@ -4,7 +4,7 @@ from india_compliance.gst_india.constants import OVERSEAS_GST_CATEGORIES
 from india_compliance.gst_india.constants.custom_fields import (
     SALES_REVERSE_CHARGE_FIELDS,
 )
-from india_compliance.gst_india.utils import toggle_custom_fields
+from india_compliance.gst_india.utils.custom_fields import toggle_custom_fields
 
 # Enable setting only if transaction exists in last 3 years.
 POSTING_DATE_CONDITION = {
