+++ india_compliance/patches/post_install/update_e_invoice_fields_and_logs.py
@@ -1,6 +1,7 @@
 import frappe
 
 from india_compliance.gst_india.utils import parse_datetime
+from india_compliance.gst_india.utils.custom_fields import delete_custom_fields
 
 user = None
 
@@ -317,32 +318,3 @@ def delete_e_invoice_fields():
         ]
     }
     delete_custom_fields(FIELDS_TO_DELETE)
-
-
-### Helper Function
-
-
-def delete_custom_fields(custom_fields):
-    """
-    :param custom_fields: a dict like `{'Sales Invoice': [{fieldname: 'test', ...}]}`
-    """
-
-    for doctypes, fields in custom_fields.items():
-        if isinstance(fields, dict):
-            # only one field
-            fields = [fields]
-
-        if isinstance(doctypes, str):
-            # only one doctype
-            doctypes = (doctypes,)
-
-        for doctype in doctypes:
-            frappe.db.delete(
-                "Custom Field",
-                {
-                    "fieldname": ("in", [field["fieldname"] for field in fields]),
-                    "dt": doctype,
-                },
-            )
-
-            frappe.clear_cache(doctype=doctype)
