+++ india_compliance/patches/post_install/update_reverse_charge_and_export_type.py
@@ -1,6 +1,6 @@
 import frappe
 
-from india_compliance.gst_india.utils import delete_old_fields
+from india_compliance.gst_india.utils.custom_fields import delete_old_fields
 
 DOCTYPES = ("Purchase Invoice", "Sales Invoice")
