+++ india_compliance/patches/v14/remove_ecommerce_gstin_from_purchase_invoice.py
@@ -1,4 +1,4 @@
-from india_compliance.gst_india.utils import delete_old_fields
+from india_compliance.gst_india.utils.custom_fields import delete_old_fields
 
 
 def execute():
